requirements.txt
selenium==4.0.0
webdriver-manager==4.0.1
PyYAML==6.0.1